+++ packages/authhero/CHANGELOG.md
@@ -1,5 +1,16 @@
 # authhero
 
+## 0.187.0
+
+### Minor Changes
+
+- Add resource servers, rules and permissions
+
+### Patch Changes
+
+- Updated dependencies
+  - @authhero/adapter-interfaces@0.79.0
+
 ## 0.186.0
 
 ### Minor Changes